+++ TaxCountry/dingxiangTax_C/1OLD.txt
@@ -1,0 +1,43 @@
+39.108.173.138:80
+114.215.102.168:8081
+202.109.207.125:8888
+153.3.235.82:80
+163.177.151.23:80
+119.39.74.213:80
+180.149.131.67:80
+121.8.98.198:80
+27.221.93.217:80
+175.147.96.73:8080
+222.189.127.111:9999
+113.76.96.45:9797
+183.51.191.183:9797
+202.117.120.242:8080
+119.49.144.83:80
+113.209.78.181:8998
+183.146.211.142:8080
+218.28.58.150:53281
+183.30.197.253:9797
+218.56.132.157:8080
+111.13.111.184:80
+120.78.78.141:8888
+113.65.20.75:9999
+113.99.218.73:9797
+125.66.140.27:53281
+123.125.142.40:80
+111.2.122.46:8080
+121.232.194.33:9000
+121.42.224.91:80
+116.11.254.37:80
+112.80.255.21:80
+123.156.103.46:8080
+121.8.98.196:80
+202.100.83.139:80
+119.249.48.235:80
+123.139.56.238:9999
+27.46.42.235:9797
+42.224.219.103:8118
+58.247.127.145:53281
+112.91.218.21:9000
+118.114.77.47:8080
+113.79.74.109:9797
+122.136.212.132:53281
