--- TaxCountry/dingxiangTax_C/ghostdriver.log
@@ -1,1 +1,0 @@
-[INFO  - 2018-01-11T11:22:32.122Z] GhostDriver - Main - running on port 50561
